meson: add dependency override for libslirp
If libslirp is built as a Meson subproject, Meson needs to associate the
superproject's dependency('slirp') call with our libslirp_libdep variable.
The slirp.wrap file shipped in Meson wrapdb accomplishes this with:

    [provide]
    slirp = libslirp_libdep

But that requires wrapdb to know the name of our libslirp_libdep variable,
which changed in 04612eb31674.

Explicitly bind the pkg-config name 'slirp' to libslirp_libdep.  This will
allow slirp.wrap to say:

    [provide]
    dependency_names = slirp

and avoid hardcoding the variable name.