apple: Fix getting IPv4 DNS server address when IPv4 and IPv4 are interleaved
When getting an IPv4 DNS server address, if libresolv returns

IPv4
IPv6
IPv4
IPv6

(or just IPv4 and IPv6)

we would still have found == 1 on the second iteration and thus take the
IPv6 even if it's not the proper af. We can as well just completely ignore
the non-matching af entries.

Fixes #85