tcp: Fix starting the linger2 timer on socket shutdown
If the socket is completely shut down just before we get our FIN ack-ed, the
socket state has SS_NOFDREF instead of SS_FCANTRCVMORE.

If the socket sees write() shut down (and we thus switch to FIN_WAIT_1), and
we get our FIN acked before seeing read() shut down, when we eventually see
that we need to start the linger2 timer.

This way, it's only if we still have a live Internet socket that we stay in
the TCPS_FIN_WAIT_2 state without a timer, and we let the host decide the
linger2 timer.

Fixes #86

(hopefully)