Move the modified 3-Clause BSD text into LICENSE
This is to help the automated license detectors find the text without
getting confused by the surrounding verbiage. However the text has
is lightly modified from the reference 3-clause text I checked on:

  https://opensource.org/license/bsd-3-clause

As it adds:

  "All rights reserved" on the copyright line
  Mentions DANNY GASPAROVSKI in place of COPYRIGHT HOLDERS
  Uses `` and '' to wrap "AS IS"

Which results in the golang license detector only having 96%
confidence this is a BSD-3-Clause. Before this patch it was 89%
confident. However these seems enough to convince GitLab to properly
categorise the license of the project.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>